+++ src/components/DetailFooter.js
